feat: daily product card와 list 분리
