+++ src/_shared/modules/deal/components/DailyProductList.tsx
@@ -1,45 +1,10 @@
-import Link from "next/link";
-
 import { ProductPreviewInfo } from "@/lib/types/productType";
 
-import { OPTION_TEXT, TITLE } from "@/lib/styles";
-
-import { DealTimer } from "@/_shared/modules/deal/components/DealTimer";
-import { ProductImage } from "@/_shared/modules/product/components/ProductImage";
-import { ProductPrice } from "@/_shared/modules/product/components/ProductPrice";
-
+import { DailyProductCard } from "./DailyProductCard";
 import { fetchProductPreviewInfo } from "@/lib/api/productsApi";
+
 interface ProductCardProps {
   product: ProductPreviewInfo;
-}
-
-function DailyProductCard({ product }: ProductCardProps) {
-  return (
-    <li>
-      <Link href={`/product/${product.id}`}>
-        <div className="flex-shrink-0 w-full rounded-lg overflow-hidden">
-          <div className="relative h-48 w-full rounded-t-lg overflow-hidden">
-            <ProductImage product={product} />
-          </div>
-          <div className="p-4">
-            <h3 className={TITLE}>{product.name}</h3>
-            <div className="flex items-center justify-between">
-              <ProductPrice product={product} size="large" />
-              <div className="flex items-center">
-                <span className="text-yellow-400">★</span>
-                <span className={OPTION_TEXT}>
-                  {product.averageRating} ({product.reviewCount})
-                </span>
-              </div>
-            </div>
-            <div>
-              <DealTimer />
-            </div>
-          </div>
-        </div>
-      </Link>
-    </li>
-  );
 }
 
 async function DailyProductList() {
